requirements.txt
langchain>=0.1.0
langchain-community>=0.0.16
langchain-core>=0.1.10
langchain-groq>=0.0.1
langchain-huggingface>=0.0.1
langgraph>=0.0.15
pydantic>=2.0.0
sentence-transformers>=2.2.2
scikit-learn>=1.0.2
numpy>=1.21.0
python-dotenv>=1.0.0
pypdf>=3.0.0
fastapi>=0.104.0
uvicorn>=0.24.0 